+++ cellpylib/langtons_loop.py
@@ -257,7 +257,7 @@ class LangtonsLoop(CTRBLRule):
         :return: the initial conditions
         """
 
-        initial_conditions = np.zeros(dim, dtype=np.int)
+        initial_conditions = np.zeros(dim, dtype=np.int32)
         for i in range(n):
             row_i = row[i]
             col_i = col[i]
